asklio-procurement-backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg[binary]==3.2.10
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
weaviate-client==4.8.1
httpx==0.27.0
tenacity==8.3.0
structlog==24.1.0
python-dotenv==1.0.1

passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0